mmr-testing/MMR-Source-url-set-to-childNode.html

<!DOCTYPE html>
<html>
<head>
    <title>Source is set to ChildNode test case.</title>
    <meta name="timeout" content="normal">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="mediasource-util.js"></script>
</head>
<body>
    <div id="log"></div>
    <script>
        mediasource_test(function (test) {

            var mediaElement = document.createElement("video");
            if (!document.body) {
                document.body = document.createElement("body");
            }
            
            var sourceMP4 = document.createElement("source");
            sourceMP4.type = 'video/mp4; codecs="mp4a.40.2, avc1.4d400d"';
            sourceMP4.src = "mp4/test.mp4";
            mediaElement.appendChild(sourceMP4);
            
            //mediaElement.src = "mp4/test.mp4";
            mediaElement.autoplay = true;

            document.body.appendChild(mediaElement);

            mediaElement.addEventListener('error', test.unreached_func("Unexpected event 'error'"));
            mediaElement.addEventListener('ended', test.step_func_done());

            test.expectEvent(mediaElement, 'playing', 'Playing triggered');

            test.step_timeout(function () {
                assert_false(test.eventExpectations_.expectingEvents(), "Pending event expectations.");
            }, 5000);

            test.removeMediaElement_ = true;
            test.add_cleanup(function () {
                if (test.removeMediaElement_) {
                    document.body.removeChild(mediaTag);
                    test.removeMediaElement_ = false;
                }
            });

        }, "Test soure is set to childNode");

    </script>

</body>
</html>
